requirements.txt
tls_client
pillow
requests
colorama